requirements.txt
Django==1.7.1
argparse==1.2.1
distribute==0.6.24
django-tastypie==0.12.1
python-dateutil==2.3
python-mimeparse==0.1.4
requests==2.5.0
six==1.8.0
wsgiref==0.1.2
